refatorando forma de notificar erros na deleção de tag

+++ Features/Tags/Delete/DeleteTagHandler.cs
@@ -12,12 +12,13 @@ public sealed class DeleteTagHandler : IDeleteTagHandler
     {
         var tag = await _db.Tags.FirstOrDefaultAsync(t => t.Id == id, ct) ?? throw new KeyNotFoundException("Tag not found.");
 
-        //var inUse = await _db.NoticiasTags.AnyAsync(nt => nt.TagId == id, ct);
-        //if (inUse)
-        //{
-        //    throw new InvalidOperationException("Cannot delete this tag because it is referenced by one or more news items.");
-        //}
-            
+        var inUse = await _db.NewsTags.AnyAsync(nt => nt.TagId == id, ct);
+
+        if (inUse)
+        {
+            throw new InvalidOperationException("Cannot delete this tag because it is referenced by one or more news items.");
+        }
+
         _db.Tags.Remove(tag);
         await _db.SaveChangesAsync(ct);
     }
